+++ src/Glitch/Renderer/Html.php
@@ -95,7 +95,7 @@ class Html implements Renderer
         $output[] = $this->renderHeader('dump');
 
         $output[] = '<header class="title">';
-        $output[] = '<h1>Glitch <span class="version">'.\Glitch::VERSION.'</span></h1>';
+        $output[] = '<h1>Glitch <span class="version">'.Glitch::VERSION.'</span></h1>';
         $output[] = '</header>';
 
         $output[] = '<div class="cols">';
@@ -135,7 +135,7 @@ class Html implements Renderer
             $output[] = $this->renderProductionExceptionMessage($exception);
         } else {
             $output[] = '<header class="title">';
-            $output[] = '<h1>Glitch <span class="version">'.\Glitch::VERSION.'</span></h1>';
+            $output[] = '<h1>Glitch <span class="version">'.Glitch::VERSION.'</span></h1>';
             $output[] = '</header>';
 
             $output[] = '<div class="cols">';